优化初始化主题
