+++ lib/lang.js
@@ -2,7 +2,7 @@ import zhCN from './lang/zh-CN'
 import enUS from './lang/en-US'
 import zhHK from './lang/zh-HK'
 import zhTW from './lang/zh-TW'
-import { mergeDeep } from '@/lib/global'
+import { mergeDeep } from './utils'
 const lang = {
   'en-US': enUS,
   'zh-CN': zhCN,
