+++ lib/theme.js
@@ -1,37 +1,45 @@
 import cookie from 'react-cookies'
 import BLOG from '@/blog.config'
+import { ALL_THEME } from '@/themes'
+import { getQueryVariable } from './utils'
 
 /**
- * 初始化主题
+ * 初始化主题 , 优先级 query > cookies > systemPrefer
  * @param isDarkMode
  * @param updateDarkMode 更改主题ChangeState函数
  * @description 读取cookie中存的用户主题
  */
 export const initDarkMode = (isDarkMode, updateDarkMode) => {
-  if (typeof window !== 'undefined') {
-    if (!isDarkMode) {
-      isDarkMode = isPreferDark()
-    }
-    updateDarkMode(isDarkMode)
-    saveDarkModeToCookies(isDarkMode)
-    document.getElementsByTagName('html')[0].setAttribute('class', isDarkMode ? 'dark' : 'light')
+  const queryMode = getQueryVariable('mode')
+  if (queryMode) {
+    isDarkMode = queryMode === 'dark'
+  } else if (!isDarkMode) {
+    isDarkMode = isPreferDark()
   }
+  updateDarkMode(isDarkMode)
+  saveDarkModeToCookies(isDarkMode)
+  document.getElementsByTagName('html')[0].setAttribute('class', isDarkMode ? 'dark' : 'light')
 }
 
 /**
- * 从cookie中读取 用户默认主题
+ * 初始化主题， 优先级 query > cookies > blog.config.js
  * @param {*} theme
  * @param {*} changeTheme
  */
 export const initTheme = (theme, changeTheme) => {
-  const userTheme = loadThemeFromCookies()
-  if (userTheme !== theme) {
-    changeTheme(userTheme)
+  const queryTheme = getQueryVariable('theme')
+  if (queryTheme && ALL_THEME.indexOf(queryTheme) > -1) {
+    changeTheme(queryTheme)
+  } else {
+    const userTheme = loadThemeFromCookies()
+    if (userTheme !== theme) {
+      changeTheme(userTheme)
+    }
   }
 }
 
 /**
- * 是否优先深色模式
+ * 是否优先深色模式， 根据系统深色模式以及当前时间判断
  * @returns {*}
  */
 export function isPreferDark () {
